Updted libraries

+++ lib/ui/widgets/tape_player/models/progress_model.dart
@@ -1,5 +1,5 @@
 class LoadingProgressData {
-  final double percent;
+  final int percent;
   final String filePath;
 
   LoadingProgressData(this.filePath, this.percent);
